requirements.txt
asttokens==2.4.1
black==24.1a1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
executing==2.0.1
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
ipython==8.19.0
jedi==0.19.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyfiglet==1.0.2
Pygments==2.17.2
pyparsing==3.1.1
pyright==1.1.343
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.14.0
typer==0.9.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.12
